Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.7.4,<3.0.0
httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.7
groq==0.11.0
python-dotenv==1.0.1
langgraph==0.1.19
langchain-core==0.2.23
sqlalchemy==2.0.32
